Handle matching against multiple stack versions
As encountered in #58, if serverless-haskell gets pinned via extra-deps in a stack.yaml file, it will result in 
multiple versions being available. Before the two version lines would get matched against. This splits the 
version output on newlines, checks if the `javascriptPackageVersion` is in the array, and fails if not,
 joining back the haskell versions with commas.