Fix format of SNS events
Looks like some of the fields coming from SNS don't match the names used in the Haskell code. Also MessageAttributes appears to be optional, so it is now a Maybe - this is also why `omitNothingFields` has also been added to the FromJSON instance. This is untested, I can try it and check it works on Monday.